requirements.txt
# Core Dependencies
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.12.1

# FastAPI & Web Framework
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
python-multipart==0.0.21
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0

# Gradio UI
gradio==6.3.0
gradio_client==2.0.3

# LangChain & RAG
langchain==1.2.3
langchain-core==1.2.7
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langsmith==0.6.2

# Vector Database
qdrant-client==1.16.2

# Embeddings & ML
sentence-transformers==5.2.0
transformers==4.57.3
tokenizers==0.22.2
torch==2.2.2
safetensors==0.7.0
huggingface-hub==0.36.0
fastembed==0.7.4

# Scientific Computing
numpy==2.4.1
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.17.0

# Document Processing
PyMuPDF==1.26.6
pymupdf4llm==0.2.9
python-docx==1.1.2

# OSS Storage
oss2==2.18.4

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
tenacity==9.1.2
tqdm==4.67.1
click==8.3.1
typer==0.21.1
rich==14.2.0
loguru==0.7.3

# Jupyter (Optional - for notebook usage)
ipykernel==7.1.0
ipython==9.9.0
jupyter_client==8.8.0
jupyter_core==5.9.1
